I want to pull request into my FORK this time

+++ Java/Calculator/Calculator.java
@@ -1,10 +1,10 @@
 package com.company;
 
 public class Calculator {
-    private int number1, number2;
+    private double number1, number2;
     private double result;
 
-    public Calculator(int number1, int number2) {
+    public Calculator(double number1, double number2) {
         this.number1 = number1;
         this.number2 = number2;
     }
@@ -12,27 +12,27 @@ public class Calculator {
     public Calculator() {
     }
 
-    // Methode Addition
+    // Method Addition
 //    public int addition(int number1, int number2) {
 //        return number1 + number2;
 //    }
 
-    public void addition(int number1, int number2) {
+    public void addition(double number1, double number2) {
         this.result = number1 + number2;
     }
 
-    // Methode Subtraktion
-    public void subtraction(int number1, int number2) {
+    // Method Subtraction
+    public void subtraction(double number1, double number2) {
         this.result = number1 - number2;
     }
 
-    // Methode Multiplikation
-    public int multiplication(int number1, int number2) {
+    // Method Multiplication
+    public int multiplication(double number1, double number2) {
         return number1 * number2;
     }
 
-    // Methode Diffusion
-    public double division(int number1, int number2) {
+    // Method Diffusion
+    public double division(double number1, double number2) {
         if (number2 == 0) {
             System.out.println("Fehler, durch 0 teilen nicht möglich!");
             return 0;
@@ -44,16 +44,16 @@ public class Calculator {
     public int getNumber1() {
         return number1;
     }
-
-    public void setNumber1(int number1) {
+// Inserting the number 1
+    public void setNumber1(double number1) {
         this.number1 = number1;
     }
-
+// getting number 2
     public int getNumber2() {
         return number2;
     }
 
-    public void setNumber2(int number2) {
+    public void setNumber2(double number2) {
         this.number2 = number2;
     }
